+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - <index.js>: git error notification
 - <index.js>: <.gitignore> is created only if `git` is present on the system
 - <templates.js> {gitIgnore()}: ignore `.prettierrc` too
+- <templates.js> modified the export strategy
+- <exitcodes.js> modified the export strategy
 - <index.js>: colored the terminal output
 - <templates.js>: colored the terminal output